requirements.txt
asgiref==3.2.3
Django==3.0.2
django-crispy-forms==1.9.0
gunicorn==20.0.4
pytz==2019.3
sqlparse==0.3.1
Pillow==7.0.0
django-heroku==0.3.1
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
social-auth-app-django